Flight: Clean up cruft in target code
- Move remaining hw defs into board_hw_defs.c
- Remove board.c, move the two functions into pios_board.c
