+++ flight/targets/brain/fw/pios_board.c
@@ -46,63 +46,6 @@
 #include "manualcontrolsettings.h"
 #include "onscreendisplaysettings.h"
 
-/**
- * Configuration for the MPU9250 chip
- */
-#if defined(PIOS_INCLUDE_MPU9250_BRAIN)
-#include "pios_mpu9250_brain.h"
-static const struct pios_exti_cfg pios_exti_mpu9250_cfg __exti_config = {
-	.vector = PIOS_MPU9250_IRQHandler,
-	.line = EXTI_Line13,
-	.pin = {
-		.gpio = GPIOC,
-		.init = {
-			.GPIO_Pin = GPIO_Pin_13,
-			.GPIO_Speed = GPIO_Speed_2MHz,  // XXXX
-			.GPIO_Mode = GPIO_Mode_IN,
-			.GPIO_OType = GPIO_OType_OD,
-			.GPIO_PuPd = GPIO_PuPd_NOPULL,
-		},
-	},
-	.irq = {
-		.init = {
-			.NVIC_IRQChannel = EXTI15_10_IRQn,
-			.NVIC_IRQChannelPreemptionPriority = PIOS_IRQ_PRIO_MID,
-			.NVIC_IRQChannelSubPriority = 0,
-			.NVIC_IRQChannelCmd = ENABLE,
-		},
-	},
-	.exti = {
-		.init = {
-			.EXTI_Line = EXTI_Line13, // matches above GPIO pin
-			.EXTI_Mode = EXTI_Mode_Interrupt,
-			.EXTI_Trigger = EXTI_Trigger_Rising,
-			.EXTI_LineCmd = ENABLE,
-		},
-	},
-};
-
-static const struct pios_mpu9250_cfg pios_mpu9250_cfg = {
-	.exti_cfg = &pios_exti_mpu9250_cfg,
-	.default_samplerate = 1000,  // Note: has to be generated by dividing 1kHz
-	.interrupt_cfg = PIOS_MPU60X0_INT_CLR_ANYRD,
-	.interrupt_en = PIOS_MPU60X0_INTEN_DATA_RDY,
-	.User_ctl = 0,
-	.Pwr_mgmt_clk = PIOS_MPU60X0_PWRMGMT_PLL_Z_CLK,
-	.orientation = PIOS_MPU60X0_TOP_0DEG
-};
-#endif /* PIOS_INCLUDE_MPU9250 */
-
-/**
- * Configuration for the MS5611 chip
- */
-#if defined(PIOS_INCLUDE_MS5611)
-#include "pios_ms5611_priv.h"
-static const struct pios_ms5611_cfg pios_ms5611_cfg = {
-	.oversampling = MS5611_OSR_4096,
-	.temperature_interleaving = 1,
-};
-#endif /* PIOS_INCLUDE_MS5611 */
 
 #if defined(PIOS_INCLUDE_FRSKY_RSSI)
 #include "pios_frsky_rssi_priv.h"
@@ -113,6 +56,25 @@ uintptr_t pios_com_logging_id;
 uintptr_t pios_com_openlog_logging_id;
 
 uintptr_t pios_uavo_settings_fs_id;
+
+
+/**
+ * @brief   Early initialization code.
+ * @details This initialization must be performed just after stack setup
+ *          and before any other initialization.
+ */
+void __early_init(void)
+{
+	stm32_clock_init();
+}
+
+/**
+ * @brief   Board-specific RTOS initialization code.
+ */
+void boardInit(void)
+{
+}
+
 
 /**
 * Initialise PWM Output for black/white level setting
