+++ flight/targets/brainre1/fw/pios_board.c
@@ -57,51 +57,7 @@
 #include "rgbledsettings.h"
 
 #include "pios_ir_transponder.h"
-#include "pios_ws2811.h"
 
-#if defined(PIOS_INCLUDE_BMI160)
-#include "pios_bmi160.h"
-
-static const struct pios_exti_cfg pios_exti_bmi160_cfg __exti_config = {
-	.vector = PIOS_BMI160_IRQHandler,
-	.line = EXTI_Line13,
-	.pin = {
-		.gpio = GPIOC,
-		.init = {
-			.GPIO_Pin = GPIO_Pin_13,
-			.GPIO_Speed = GPIO_Speed_2MHz,  // XXXX
-			.GPIO_Mode = GPIO_Mode_IN,
-			.GPIO_OType = GPIO_OType_OD,
-			.GPIO_PuPd = GPIO_PuPd_NOPULL,
-		},
-	},
-	.irq = {
-		.init = {
-			.NVIC_IRQChannel = EXTI15_10_IRQn,
-			.NVIC_IRQChannelPreemptionPriority = PIOS_IRQ_PRIO_MID,
-			.NVIC_IRQChannelSubPriority = 0,
-			.NVIC_IRQChannelCmd = ENABLE,
-		},
-	},
-	.exti = {
-		.init = {
-			.EXTI_Line = EXTI_Line13, // matches above GPIO pin
-			.EXTI_Mode = EXTI_Mode_Interrupt,
-			.EXTI_Trigger = EXTI_Trigger_Rising,
-			.EXTI_LineCmd = ENABLE,
-		},
-	},
-};
-
-static const struct pios_bmi160_cfg pios_bmi160_cfg = {
-	.exti_cfg = &pios_exti_bmi160_cfg,
-	.orientation = PIOS_BMI160_TOP_0DEG,
-	.odr = PIOS_BMI160_ODR_1600_Hz,
-	.acc_range = PIOS_BMI160_RANGE_8G,
-	.gyro_range = PIOS_BMI160_RANGE_2000DPS,
-	.temperature_interleaving = 50
-};
-#endif /* PIOS_INCLUDE_BMI160 */
 
 #if defined(PIOS_INCLUDE_FRSKY_RSSI)
 #include "pios_frsky_rssi_priv.h"
@@ -113,6 +69,23 @@ uintptr_t pios_internal_adc_id;
 uintptr_t pios_uavo_settings_fs_id;
 
 ws2811_dev_t pios_ws2811;
+
+/**
+ * @brief   Early initialization code.
+ * @details This initialization must be performed just after stack setup
+ *          and before any other initialization.
+ */
+void __early_init(void)
+{
+	stm32_clock_init();
+}
+
+/**
+ * @brief   Board-specific RTOS initialization code.
+ */
+void boardInit(void)
+{
+}
 
 /**
  * settingdUpdatedCb()
