+++ flight/targets/dtfc/fw/pios_board.c
@@ -62,6 +62,22 @@ uintptr_t pios_can_id;
 uintptr_t pios_com_openlog_logging_id;
 
 /**
+ * @brief   Early initialization code.
+ * @details This initialization must be performed just after stack setup
+ *          and before any other initialization.
+ */
+void __early_init(void) {
+
+  stm32_clock_init();
+}
+
+/**
+ * @brief   Board-specific initialization code.
+ */
+void boardInit(void) {
+}
+
+/**
  * PIOS_Board_Init()
  * initializes all the core subsystems on this specific hardware
  * called from System/openpilot.c
