+++ flight/targets/lux/board-info/board_hw_defs.c
@@ -860,6 +860,56 @@ const struct pios_usb_cdc_cfg pios_usb_cdc_cfg = {
 };
 #endif	/* PIOS_INCLUDE_USB_CDC */
 
+
+/**
+ * Configuration for the MPU9250 chip
+ */
+#if defined(PIOS_INCLUDE_MPU9250_SPI)
+#include "pios_mpu9250.h"
+static const struct pios_exti_cfg pios_exti_mpu9250_cfg __exti_config = {
+    .vector = PIOS_MPU9250_IRQHandler,
+    .line = EXTI_Line5,
+    .pin = {
+        .gpio = GPIOA,
+        .init = {
+            .GPIO_Pin = GPIO_Pin_5,
+            .GPIO_Speed = GPIO_Speed_50MHz,
+            .GPIO_Mode = GPIO_Mode_IN,
+            .GPIO_OType = GPIO_OType_OD,
+            .GPIO_PuPd = GPIO_PuPd_NOPULL,
+        },
+    },
+    .irq = {
+        .init = {
+            .NVIC_IRQChannel = EXTI9_5_IRQn,
+            .NVIC_IRQChannelPreemptionPriority = PIOS_IRQ_PRIO_HIGH,
+            .NVIC_IRQChannelSubPriority = 0,
+            .NVIC_IRQChannelCmd = ENABLE,
+        },
+    },
+    .exti = {
+        .init = {
+            .EXTI_Line = EXTI_Line5, // matches above GPIO pin
+            .EXTI_Mode = EXTI_Mode_Interrupt,
+            .EXTI_Trigger = EXTI_Trigger_Rising,
+            .EXTI_LineCmd = ENABLE,
+        },
+    },
+};
+
+static struct pios_mpu9250_cfg pios_mpu9250_cfg = {
+    .exti_cfg = &pios_exti_mpu9250_cfg,
+    .default_samplerate = 1000,
+    .interrupt_cfg = PIOS_MPU60X0_INT_CLR_ANYRD,
+
+    .use_magnetometer = false,
+    .default_gyro_filter = PIOS_MPU9250_GYRO_LOWPASS_184_HZ,
+    .default_accel_filter = PIOS_MPU9250_ACCEL_LOWPASS_184_HZ,
+    .orientation = PIOS_MPU9250_TOP_180DEG
+};
+
+#endif /* PIOS_INCLUDE_MPU9250_SPI */
+
 /**
  * @}
  * @}
