+++ flight/targets/revolution/fw/Makefile
@@ -103,7 +103,6 @@ SRC += ${foreach MOD, ${MODULES} ${OPTMODULES}, ${wildcard ${OPMODULEDIR}/${MOD}
 SRC += main.c
 SRC += pios_board.c
 SRC += pios_usb_board_data.c
-SRC += board.c
 SRC += $(FLIGHTLIB)/alarms.c
 SRC += $(OPUAVTALK)/uavtalk.c
 SRC += $(OPUAVOBJ)/uavobjectmanager.c
